Project/engine/pyproject.toml
[project]
name = "engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.37.1",
    "librosa>=0.11.0",
    "openai-whisper>=20250625",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "spikingjelly>=0.0.0.0.14",
]
